Organizes files.
